+++ test/test_sync_upload.py
@@ -31,15 +31,15 @@ class UploadResolveTest(_SyncTestBase):
         opts = {"verbose": self.verbose}  # default options, i.e. 'skip' conflicts
         # Default options: expect 4 unresolved conflicts
         stats = self.do_run_suite(UploadSynchronizer, opts)
-        # self._dump_de_facto_results(stats)
+        self._dump_de_facto_results(stats)
 
         # We expect 7 conflicts, and leave them unresolved (i.e. skip them all)
 
         self.assertEqual(stats["files_written"], 3)
         self.assertEqual(stats["download_files_written"], 0)
         self.assertEqual(stats["upload_files_written"], 3)
-        self.assertEqual(stats["files_deleted"], 1)
-        self.assertEqual(stats["dirs_deleted"], 2)
+        self.assertEqual(stats["files_deleted"], 0)
+        self.assertEqual(stats["dirs_deleted"], 0)
         self.assertEqual(stats["conflict_files"], 7)
         self.assertEqual(stats["conflict_files_skipped"], 7)
 
@@ -51,6 +51,7 @@ class UploadResolveTest(_SyncTestBase):
         expect_remote = {
             "file1.txt": {"content": "local1", "date": "2014-01-01 12:00:00"},
             "file2.txt": {"content": "local 13:00", "date": "2014-01-01 13:00:00"},
+            "file3.txt": {"content": "local3", "date": "2014-01-01 12:00:00"},
             "file4.txt": {"content": "remote 13:00", "date": "2014-01-01 13:00:00"},
             "file6.txt": {"content": "remote 13:00:05", "date": "2014-01-01 13:00:05"},
             "file7.txt": {"content": "remote 13:00", "date": "2014-01-01 13:00:00"},
@@ -61,6 +62,14 @@ class UploadResolveTest(_SyncTestBase):
             },
             "folder2/file2_1.txt": {
                 "content": "local 13:00",
+                "date": "2014-01-01 13:00:00",
+            },
+            "folder3/file3_1.txt": {
+                "content": "local3_1",
+                "date": "2014-01-01 12:00:00",
+            },
+            "folder4/file4_1.txt": {
+                "content": "remote 13:00",
                 "date": "2014-01-01 13:00:00",
             },
             "folder5/file5_1.txt": {
@@ -172,6 +181,60 @@ class UploadResolveTest(_SyncTestBase):
                 "date": "2014-01-01 13:00:00",
             },
             "new_file1.txt": {"content": "local 13:00", "date": "2014-01-01 13:00:00"},
+        }
+        self.assert_test_folder_equal(get_test_folder("remote"), expect_remote)
+
+    def test_delete(self):
+        opts = {"verbose": self.verbose, "resolve": "skip", "delete": True}
+
+        stats = self.do_run_suite(UploadSynchronizer, opts)
+        self._dump_de_facto_results(stats)
+
+        # We expect 7 conflicts, and leave them unresolved (i.e. skip them all)
+
+        self.assertEqual(stats["files_written"], 3)
+        self.assertEqual(stats["download_files_written"], 0)
+        self.assertEqual(stats["upload_files_written"], 3)
+        self.assertEqual(stats["files_deleted"], 2)
+        self.assertEqual(stats["dirs_deleted"], 2)
+        self.assertEqual(stats["conflict_files"], 7)
+        self.assertEqual(stats["conflict_files_skipped"], 7)
+
+        self.assert_test_folder_equal(
+            get_test_folder("local"), _SyncTestBase.local_fixture_modified
+        )
+
+        # We expect that remote only contains files that match '*1.txt'
+        expect_remote = {
+            "file1.txt": {"content": "local1", "date": "2014-01-01 12:00:00"},
+            "file2.txt": {"content": "local 13:00", "date": "2014-01-01 13:00:00"},
+            "file4.txt": {"content": "remote 13:00", "date": "2014-01-01 13:00:00"},
+            "file6.txt": {"content": "remote 13:00:05", "date": "2014-01-01 13:00:05"},
+            "file7.txt": {"content": "remote 13:00", "date": "2014-01-01 13:00:00"},
+            "file8.txt": {"content": "remote 13:00", "date": "2014-01-01 13:00:00"},
+            "folder1/file1_1.txt": {
+                "content": "local1_1",
+                "date": "2014-01-01 12:00:00",
+            },
+            "folder2/file2_1.txt": {
+                "content": "local 13:00",
+                "date": "2014-01-01 13:00:00",
+            },
+            "folder5/file5_1.txt": {
+                "content": "remote 13:00",
+                "date": "2014-01-01 13:00:00",
+            },
+            "new_file1.txt": {"content": "local 13:00", "date": "2014-01-01 13:00:00"},
+            "new_file3.txt": {"content": "local 13:00", "date": "2014-01-01 13:00:00"},
+            "new_file4.txt": {
+                "content": "remote 13:00 with other content",
+                "date": "2014-01-01 13:00:00",
+            },
+            "new_file5.txt": {
+                "content": "remote 13:00:05",
+                "date": "2014-01-01 13:00:05",
+            },
+            "new_file6.txt": {"content": "remote 13:00", "date": "2014-01-01 13:00:00"},
         }
         self.assert_test_folder_equal(get_test_folder("remote"), expect_remote)